+++ src/components/ReaderLayout/ReaderLayout.tsx
@@ -26,6 +26,7 @@ import Link from "../Link/Link";
 import { addTocToManifest } from "@nypl/web-reader";
 import Loading from "../Loading/Loading";
 import { trackCtaClick } from "~/src/lib/adobe/Analytics";
+import NotFound404 from "~/src/pages/404";
 
 const origin =
   typeof window !== "undefined" && window.location?.origin
@@ -147,6 +148,10 @@ const ReaderLayout: React.FC<{
       </Link>
     );
   };
+
+  if (!isEmbed && !isRead) {
+    return NotFound404();
+  }
 
   return (
     <>