+++ src/types/DataModel.ts
@@ -65,6 +65,7 @@ export type LinkFlags = {
   edd?: boolean;
   embed?: boolean;
   nypl_login?: boolean;
+  fulfill_limited_access?: boolean;
 };
 
 export type ItemLink = {