+++ src/util/EditionCardUtils.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import {
   Agent,
   Instance,
@@ -16,6 +15,8 @@ import {
   PLACEHOLDER_COVER_LINK,
 } from "../constants/editioncard";
 import { MediaTypes } from "../constants/mediaTypes";
+import { NextRouter } from "next/router";
+import { LOGIN_LINK_BASE } from "../constants/links";
 
 // EditionCard holds all the methods needed to build an Edition Card
 export default class EditionCardUtils {
@@ -220,4 +221,41 @@ export default class EditionCardUtils {
         : undefined;
     return universityPress !== undefined;
   }
+
+  static createGetContent = (nyplIdentityCookie: any, router: NextRouter) => {
+    const fetchWithAuth = async (fulfillUrl: string, proxyUrl?: string) => {
+      const url = new URL(fulfillUrl);
+      const res = await fetch(url.toString(), {
+        method: "GET",
+        headers: {
+          Authorization: `Bearer ${nyplIdentityCookie.access_token}`,
+        },
+      });
+
+      if (res.ok) {
+        // Generate the resource URL using the proxy
+        const resourceUrl = res.url;
+        const proxiedUrl: string = proxyUrl
+          ? `${proxyUrl}${encodeURIComponent(resourceUrl)}`
+          : resourceUrl;
+        const response = await fetch(proxiedUrl, { mode: "cors" });
+        const resourceAsByteArray = new Uint8Array(
+          await response.arrayBuffer()
+        );
+
+        if (!response.ok) {
+          throw new Error("Response not Ok for URL: " + url);
+        }
+        return resourceAsByteArray;
+      } else {
+        // redirect to the NYPL login page if access token is invalid
+        if (res.status === 401) {
+          router.push(
+            LOGIN_LINK_BASE + encodeURIComponent(window.location.href)
+          );
+        }
+      }
+    };
+    return fetchWithAuth;
+  };
 }