+++ playwright/support/mappings.ts
@@ -159,7 +159,8 @@ export const elements = {
   "catalog header link": "span:text('Go To The Catalog')",
   "research catalog header link": "span:text('Go To The Research Catalog')",
   "close my account header link": "[aria-label='Close']",
-  "locations header link": "[href='https://www.nypl.org/locations']",
+  "locations header link":
+    "[href='https://www.nypl.org/locations']:text('Locations')",
   "library card header link": "[href='https://www.nypl.org/library-card/new']",
   "email updates header link":
     "[href='https://pub.email.nypl.org/subscriptioncenter']",
@@ -203,6 +204,25 @@ export const elements = {
     "//a[@href='http://www.nypl.org/help/about-nypl/legal-notices/rules-and-regulations']",
   "about NYPL footer link": "//a[@href='http://www.nypl.org/help/about-nypl']",
   "language footer link": "//a[@href='http://www.nypl.org/language']",
+
+  /** license page locators */
+  "license explanations header": "//h1[text()='License Explanations']",
+  "public domain header": "//h2[text()='Public Domain']",
+  "creative commons licenses header":
+    "//h2[text()='Creative Commons Licenses']",
+  "public domain us only header": "//h2[text()='Public Domain (US Only)']",
+  "public domain subheader":
+    "//p[contains(text(),'Works may be in the public domain in the Unites St')]",
+  "public domain us only subheader":
+    "//p[contains(text(),'Works may be in the public domain in the Unites States')]",
+
+  /** external page locators */
+  "locations page header":
+    "//div[contains(text(),'Welcome to The New York Public Library. Discover o')]",
+  "get a library card page header": "//h1[@id='hero-banner']",
+  "get email updates page header": "//h1[@id='page-title']",
+  "donation page button": "//button[text()='Donate']",
+  "shop page footer": "//h2[text()='All Proceeds Support the Library']",
 };
 
 export const inputs = {