+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # CHANGE LOG
 
 ## [Pre-release]
+
 - Fix cut off text on search bar dropdown
 - Fix broken link on the About page
+- SFR-2008: Automate License Page Headers and Sub-Headers
 
 ## [0.18.1]
 
@@ -13,7 +15,7 @@
 - Update README to include production release information
 - Upgrade NYPL Design System to 3.1.1
 - Upgrade Web Reader to 4.3.4
-- Fix issue where 0 character is shown when there are no authors 
+- Fix issue where 0 character is shown when there are no authors
 
 ## [0.18.0]