+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Fix cut off text on search bar dropdown
 - Fix broken link on the About page
 - SFR-2008: Automate License Page Headers and Sub-Headers
+- SFR-2033: Verify the external NYPL header links of DRB App
+- Update PR template with new Jira link
 - Add error page for /read links with invalid source
 
 ## [0.18.1]