+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - SFR-2008: Automate License Page Headers and Sub-Headers
 - SFR-2033: Verify the external NYPL header links of DRB App
 - Update PR template with new Jira link
+- Add error page for /read links with invalid source
 
 ## [0.18.1]