+++ CHANGELOG.md
@@ -6,7 +6,11 @@
 - Fix broken link on the About page
 - SFR-2008: Automate License Page Headers and Sub-Headers
 - SFR-2033: Verify the external NYPL header links of DRB App
+- Update PR template with new Jira link
+- Add error page for /read links with invalid source
+- Implement "Read Online" for UP items
 - SFR-2032: Update local API url
+
 
 ## [0.18.1]